Fix:#27-file structure fixing and add slider.
