--- artlift_api/artlift-frontend/src/app/components/features/ArtworkThumbnail.tsx
@@ -1,17 +1,0 @@
-import { Artwork } from "@/app/data/artists";
-
-type Props = {
-  artworkThumbnail: string;
-};
-
-export default function ArtworkThumbnail({ artworkThumbnail }: Props) {
-  return (
-    <>
-        <img
-          className="object-cover w-full h-full"
-          src={artworkThumbnail}
-          alt={`Thumbnail ${artworkThumbnail}`}
-        />
-    </>
-  );
-}
